requirements.txt
Django==1.11.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
markdown2==2.3.5
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2017.3
six==1.11.0
